Change getter name to is because csharp wrapper remove the 'get' token from...

See merge request BC/public/liblinphone!1285

+++ tester/account_creator_xmlrpc_tester.c
@@ -48,7 +48,7 @@ static void account_creator_cb(LinphoneAccountCreator *creator, LinphoneAccountC
 	LinphoneAccountCreatorCbs *cbs = linphone_account_creator_get_current_callbacks(creator);
 
 	LinphoneAccountCreatorStatus expected_status = (LinphoneAccountCreatorStatus)(
-		(long)linphone_account_creator_service_get_user_data(linphone_account_creator_get_service(creator))
+		(intptr_t)linphone_account_creator_service_get_user_data(linphone_account_creator_get_service(creator))
 	);
 	BC_ASSERT_EQUAL(
 		status,
@@ -1235,7 +1235,7 @@ static void server_update_account_password_for_non_existent_account(void) {
 static void login_linphone_account_creator_cb(LinphoneAccountCreator *creator, LinphoneAccountCreatorStatus status, const char* resp) {
 	LinphoneAccountCreatorCbs *cbs = linphone_account_creator_get_current_callbacks(creator);
 	LinphoneAccountCreatorStatus expected_status = (LinphoneAccountCreatorStatus)(
-		(long)linphone_account_creator_service_get_user_data(linphone_account_creator_get_service(creator))
+		(intptr_t)linphone_account_creator_service_get_user_data(linphone_account_creator_get_service(creator))
 	);
 	BC_ASSERT_EQUAL(
 		status,
